modify checkout product

+++ src/main/java/com/webtrucking/dao/ProductDAO.java
@@ -3,5 +3,5 @@ package com.webtrucking.dao;
 import com.webtrucking.entity.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductDAO extends CrudRepository<Product, Long> {
+public interface ProductDAO extends CrudRepository<Product, Integer> {
 }
